// src/app/model/Graph.ts
/**
 * UnBlock RESTful Web API
 * An Bitcoin analysis service
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Graph {
    addressNodes?: Array<models.Address>;

    beneficiaryBankEdges?: Array<models.BeneficiaryBankEdge>;

    blockNodes?: Array<models.Block>;

    companyNodes?: Array<models.Company>;

    ethAddressNodes?: Array<models.EthAddress>;

    ethBlockNodes?: Array<models.EthBlock>;

    ethHasTxEdges?: Array<models.EthHasTxEdge>;

    ethMinedByEdges?: Array<models.EthMinedByEdge>;

    ethNextBlockEdges?: Array<models.EthNextBlockEdge>;

    ethOwnEdges?: Array<models.EthOwnEdge>;

    ethReceiveEdges?: Array<models.EthReceiveEdge>;

    ethRemitEdges?: Array<models.EthRemitEdge>;

    ethTxNodes?: Array<models.EthTx>;

    hasTxEdges?: Array<models.HasTxEdge>;

    interBankEdges?: Array<models.InterBankEdge>;

    managedByEdges?: Array<models.ManagedByEdge>;

    nextBlockEdges?: Array<models.NextBlockEdge>;

    ownEdges?: Array<models.OwnEdge>;

    personNodes?: Array<models.Person>;

    remitEdges?: Array<models.RemitEdge>;

    srcBankEdges?: Array<models.SrcBankEdge>;

    toBeneficiaryEdges?: Array<models.ToBeneficiaryEdge>;

    transferNodes?: Array<models.Transfer>;

    txNodes?: Array<models.Tx>;

    vinEdges?: Array<models.VinEdge>;

    voutEdges?: Array<models.VoutEdge>;

    voutNodes?: Array<models.Vout>;

}