// src/app/model/Block.ts
/**
 * UnBlock RESTful Web API
 * An Bitcoin analysis service
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Block {
    difficulty?: number;

    hash?: string;

    height?: number;

    id?: number;

    idObject?: models.IdObject;

    merkleRoot?: string;

    nonce?: number;

    size?: number;

    time?: number;

    txs?: Array<models.Tx>;

    type?: string;

    version?: number;

}