// src/app/model/Person.ts
/**
 * UnBlock RESTful Web API
 * An Bitcoin analysis service
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Person {
    alias?: string;

    birthDate?: number;

    birthPlace?: string;

    fullName?: string;

    id?: number;

    idObject?: models.IdObject;

    mainCountry?: string;

    position?: string;

    sanctioned?: boolean;

    type?: string;

}